Move operation traits into own mod, update docs
